+++ frontend/openchat-agent/src/services/openchatAgent.ts
@@ -3372,23 +3372,16 @@ export class OpenChatAgent extends EventTarget {
     async claimPrize(
         chatId: MultiUserChatIdentifier,
         messageId: bigint,
+        signInProof: string | undefined,
     ): Promise<ClaimPrizeResponse> {
         if (offline()) return Promise.resolve(CommonResponses.offline());
 
-        switch (chatId.kind) {
-            case "group_chat": {
-                const localUserIndex = await this.getGroupClient(chatId.groupId).localUserIndex();
-                const localUserIndexClient = this.getLocalUserIndexClient(localUserIndex);
-                return localUserIndexClient.claimPrize(chatId, messageId);
-            }
-            case "channel": {
-                const localUserIndex = await this.communityClient(
-                    chatId.communityId,
-                ).localUserIndex();
-                const localUserIndexClient = this.getLocalUserIndexClient(localUserIndex);
-                return localUserIndexClient.claimPrize(chatId, messageId);
-            }
-        }
+        const localUserIndex = await (chatId.kind === "group_chat"
+            ? this.getGroupClient(chatId.groupId).localUserIndex()
+            : this.communityClient(chatId.communityId).localUserIndex());
+
+        const localUserIndexClient = this.getLocalUserIndexClient(localUserIndex);
+        return localUserIndexClient.claimPrize(chatId, messageId, signInProof);
     }
 
     payForDiamondMembership(