+++ frontend/openchat-client/src/openchat.ts
@@ -6635,12 +6635,13 @@ export class OpenChat {
         chatId: MultiUserChatIdentifier,
         messageId: bigint,
         e: MouseEvent,
+        signInProof: string | undefined,
     ): Promise<boolean> {
         if (!this.#validMouseEvent(e)) {
             return Promise.resolve(false);
         }
 
-        return this.#sendRequest({ kind: "claimPrize", chatId, messageId })
+        return this.#sendRequest({ kind: "claimPrize", chatId, messageId, signInProof })
             .then((resp) => {
                 if (resp.kind !== "success") {
                     return false;
@@ -7970,6 +7971,14 @@ export class OpenChat {
         return this.#sendRequest({
             kind: "setCachedWebAuthnKey",
             key,
+        });
+    }
+
+    getSignInProof(identityKey: ECDSAKeyIdentity, delegation: DelegationChain): Promise<string> {
+        return this.#sendRequest({
+            kind: "getSignInProof",
+            identityKey: identityKey.getKeyPair(),
+            delegation: delegation.toJSON(),
         });
     }