+++ frontend/openchat-shared/src/domain/worker.ts
@@ -471,6 +471,7 @@ export type WorkerRequest =
     | ReinstateMissedDailyClaims
     | VerifyAccountLinkingCode
     | FinaliseAccountLinkingWithCode
+    | GetSignInProof
     | PayForPremiumItem
     | SetPremiumItemCost
     | OneSecEnableForwarding
@@ -1690,6 +1691,12 @@ type FinaliseAccountLinkingWithCode = {
     webAuthnKey?: WebAuthnKeyFull;
 };
 
+type GetSignInProof = {
+    kind: "getSignInProof";
+    identityKey: CryptoKeyPair;
+    delegation: JsonnableDelegationChain;
+};
+
 /**
  * Worker error type
  */
@@ -1918,6 +1925,7 @@ type DeleteFailedMessage = {
 type ClaimPrize = {
     chatId: MultiUserChatIdentifier;
     messageId: bigint;
+    signInProof: string | undefined;
     kind: "claimPrize";
 };
 
@@ -2632,6 +2640,8 @@ export type WorkerResult<T> = T extends Init
     ? void
     : T extends CreateAccountLinkingCode
     ? AccountLinkingCode | undefined
+    : T extends GetSignInProof
+    ? string
     : T extends ReinstateMissedDailyClaims
     ? boolean
     : T extends OneSecEnableForwarding